Moved GetValue method body to the Property object to allow easy conversion of an untyped property without having to go via the fieldset again.
